requirements.txt
streamlit
numpy
transformers
sentence-transformers
SpeechRecognition
pyaudio
gTTS
pydub
langdetect
deep-translator
faiss-cpu
